+++ networks/detector.py
