+++ utilities/counting/bb_polygon.py
