+++ utilities/tools/cocosplit.py
