+++ utilities/tools/split_images.py
